Create maze.java
This Java program generates a random maze and solves it by displaying it in the console. The maze is created using a depth-first search algorithm. It starts with a grid of walls and carves a path through it to create a maze. It's a fun and somewhat unconventional use of Java to create an interesting visual output in the console.